+++ packages/contracts/src/deploy-utils.ts
@@ -1,7 +1,7 @@
 import { ethers, Contract } from 'ethers'
 import { Provider } from '@ethersproject/abstract-provider'
 import { Signer } from '@ethersproject/abstract-signer'
-import { sleep, awaitCondition } from '@eth-optimism/core-utils'
+import { sleep, awaitCondition, getChainId } from '@eth-optimism/core-utils'
 import { HttpNetworkConfig } from 'hardhat/types'
 
 import { getDeployConfig } from './deploy-config'
@@ -270,8 +270,7 @@ export const getContractFromArtifact = async (
 }
 
 export const isHardhatNode = async (hre) => {
-  const { chainId } = await hre.ethers.provider.getNetwork()
-  return chainId === 31337
+  return (await getChainId(hre.ethers.provider)) === 31337
 }
 
 // Large balance to fund accounts with.