+++ packages/contracts/test/contracts/libraries/utils/Lib_BytesUtils.spec.ts
@@ -1,11 +1,8 @@
-/* Internal Imports */
-import { ethers } from 'hardhat'
 import { Contract } from 'ethers'
 
-/* External Imports */
-import { Lib_BytesUtils_TEST_JSON } from '../../../data'
-import { runJsonTest } from '../../../helpers'
 import { expect } from '../../../setup'
+import { Lib_BytesUtils_TEST_JSON } from '../../../data'
+import { deploy, runJsonTest } from '../../../helpers'
 
 describe('Lib_BytesUtils', () => {
   describe('JSON tests', () => {
@@ -15,18 +12,17 @@ describe('Lib_BytesUtils', () => {
   describe('Use of library with other memory-modifying operations', () => {
     let TestLib_BytesUtils: Contract
     before(async () => {
-      TestLib_BytesUtils = await (
-        await ethers.getContractFactory('TestLib_BytesUtils')
-      ).deploy()
+      TestLib_BytesUtils = await deploy('TestLib_BytesUtils')
     })
 
     it('should allow creation of a contract beforehand and still work', async () => {
-      const slice = await TestLib_BytesUtils.callStatic.sliceWithTaintedMemory(
-        '0x123412341234',
-        0,
-        0
-      )
-      expect(slice).to.eq('0x')
+      expect(
+        await TestLib_BytesUtils.callStatic.sliceWithTaintedMemory(
+          '0x123412341234',
+          0,
+          0
+        )
+      ).to.eq('0x')
     })
   })
 })