+++ packages/contracts/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.5.23
+
+### Patch Changes
+
+- 412688d5: Replace calls to getNetwork() with getChainId util
+
+## 0.5.22
+
+### Patch Changes
+
+- 51adb389: Add Teleportr mainnet deployment
+- Updated dependencies [5cb3a5f7]
+- Updated dependencies [6b9fc055]
+  - @eth-optimism/core-utils@0.8.4
+
 ## 0.5.21
 
 ### Patch Changes