+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @eth-optimism/sdk
 
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies [412688d5]
+  - @eth-optimism/contracts@0.5.23
+
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [51adb389]
+- Updated dependencies [5cb3a5f7]
+- Updated dependencies [6b9fc055]
+  - @eth-optimism/contracts@0.5.22
+  - @eth-optimism/core-utils@0.8.4
+
 ## 1.1.1
 
 ### Patch Changes